Convert VLCPlayer from a method to a property
VLCPlayer no longer holds a reference to the currently playing Medium. libvlc holds an internal reference so we don't need to keep ours around.

If the current medium is replaced by a call to Media.Set then libvlc will decrement the refcount. If its refcount becomes zero it will be destroyed.

Media.Get increments the refcount.